third_party/blink/web_tests/external/wpt/editing/run/inserttext_1-1000-expected.txt

This is a testharness.js-based test.
Found 9 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["inserttext","\\t"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo\\tbar" but got "foo<span style=\\"white-space:pre\\">\\t</span>bar"
[FAIL] [["defaultparagraphseparator","div"],["inserttext","\\n"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><div>bar</div>" but got "foo<div>bar</div>"
[FAIL] [["defaultparagraphseparator","p"],["inserttext","\\n"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p>" but got "foo<p>bar</p>"
[FAIL] [["defaultparagraphseparator","div"],["inserttext","abc\\ndef"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>fooabc</div><div>defbar</div>" but got "fooabc<div>defbar</div>"
[FAIL] [["defaultparagraphseparator","p"],["inserttext","abc\\ndef"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>fooabc</p><p>defbar</p>" but got "fooabc<p>defbar</p>"
[FAIL] [["inserttext"," "]] "foo []&nbsp;        bar" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "foo&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;bar" not in array ["foo&nbsp; &nbsp; bar", "foo &nbsp; &nbsp;bar"]
[FAIL] [["inserttext"," "]] " <span> </span>[]foo" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value " <span></span>&nbsp;foo" not in array [" <span> </span>&nbsp;foo", "<span></span>&nbsp;foo", "&nbsp;foo"]
[FAIL] [["inserttext"," "]] "<div style=white-space:nowrap>foo&nbsp;[]</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"white-space:nowrap\\">foo &nbsp;</div>" but got "<div style=\\"white-space:nowrap\\">foo&nbsp;&nbsp;</div>"
[FAIL] [["inserttext"," "]] "<b>a@b</b>{}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<b>a@b</b>&nbsp;" but got "<b>a@b&nbsp;</b>"
Harness: the test ran to completion.
